adding schemaTest.py for automated testing
schemaTest.py will validate a schema (e.g., smpte-tlx-items.json or smpte-tlx-profiles.json) by running a collection of test suites, each containing a number of TLX instances and an expected schema result. 

This program will require some cleanup as the directory structure is changing as I migrate the code to this repo.  Give me a minute.