added TLX validity check
added the TLX validity check to these four profile test suites.

(was already in place for ohter profile test suites.  This ensures that the test cases are "proper" TLX messages, in addition to meeting the additional profile constraints.  Otherwise, the results are confusing.

Note that in a production scenario, one would not check validity against the smpte-tlx-items schema repeatedly for each profile but instead would only do it once.