made subFrame required for ST12-3
subFrame was recognized to be REQUIRED for ST12-3 profile.

further, per ST 12-3:2016, there are only a few legal values:
00000b, 00100b, 01000b, 01100b, 10000b, and 11000b.
See ST 12-3:2016 section 6.2 describing that sub-frame_4&5 are always zero (the LSbs), sub-frame_3 must be zero unless N=5, and section 6.3 which shows the sequence of the MSbs, depending on N= 3, 4, or 5.

Note that sub-frame_1 is defined to be the MSb of the five-bit sequence, so when set is 0x10.  sub-frame_2 set give 0x08, etc.